% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterFeatures.R
\name{clusterFeatures}
\alias{clusterFeatures}
\title{Feature clustering}
\usage{
clusterFeatures(
  x,
  xin,
  rtime_var = "rtime",
  rt_cut = 10,
  cor_cut = 0.7,
  rt_grouping = c("hclust", "closest", "consecutive"),
  cor_grouping = c("louvain", "AbundanceSimilar", "none"),
  cor_use = c("everything", "all.obs", "complete.obs", "na.or.complete",
    "pairwise.complete.obs"),
  cor_method = c("pearson", "kendall", "spearman"),
  log2 = FALSE
)
}
\arguments{
\item{x}{A \linkS4class{SummarizedExperiment} object.}

\item{xin}{A string specifying which assay values to use.}

\item{rtime_var}{A string specifying the names of variable containing a
numeric vector of retention times in \code{rowData(x)}.}

\item{rt_cut}{A numeric value specifying a cut-off for the retention-time
based feature grouping.}

\item{cor_cut}{A numeric value specifying a cut-off for the
correlation-based feature grouping.}

\item{rt_grouping}{A string specifying which method to use for the
retention-time based feature grouping.}

\item{cor_grouping}{A string specifying which method to use for the
correlation-based feature grouping.}

\item{cor_use}{A string specifying which method to compute correlations in
the presence of missing values. Refer to \code{?cor} for details.}

\item{cor_method}{A string specifying which correlation coefficient is to be
computed. See \code{?cor} for details.}

\item{log2}{A logical specifying whether feature intensities needs to be
log2-transformed before calculating a correlation matrix.}
}
\value{
A \linkS4class{SummarizedExperiment} object with the grouping
results added to columns "rtime_group" (initial grouping on retention
times) and "feature_group" in its \code{rowData}.
}
\description{
Function to group LC-MS features according to their retention time and
intensity correlation across samples with a
\linkS4class{SummarizedExperiment}.
}
\details{
For soft ionization methods (e.g., LC/ESI-MS) commonly used in metabolomics,
one or more ions could be generated from an individual compound upon
ionization. The redundancy of feature data needs to be addressed since we
typically interested in compounds rather than different ion species. This
function attempts to identify a group of features from the same compound
with the following steps:
\enumerate{
\item Features are grouped by their retention times to identify co-eluting
compounds.
\item For each retention time-based group, features are further clustered by
patterns of the intensity correlations across samples to identify a subset
of features from the same compound.
}

The retention time-based grouping is performed using a hierarchical
clustering via \link{hclust} or the methods available in the \pkg{MsFeatures}
package via \link[MsFeatures:groupClosest]{MsFeatures::groupClosest} and \link[MsFeatures:groupConsecutive]{MsFeatures::groupConsecutive}.
For the \code{rt_grouping} = "hclust", complete-linkage clustering is
conducted using the Manhattan distance (i.e., difference in retention times)
where the distance between two clusters is defined as the difference in
retention times between the farthest pair of elements in the two clusters.
Group memberships are assigned by specifying the cut height for the distance
metric. For the "closest" and "consecutive", please refer to
\code{?MsFeatures::groupClosest} and \code{?MsFeatures::groupConsecutive}
for details.

For the correlation-based grouping, \code{cor_grouping} = "louvain" creates
a undirected graph using feature correlations as an adjacency matrix.
Correlations below the cut-off specified by \code{cor_cut} will be ignored. The
Louvain algorithm is used to group features via \link[igraph:cluster_louvain]{igraph::cluster_louvain}.
For the "AbundanceSimilar", \link[MsFeatures:groupSimilarityMatrix]{MsFeatures::groupSimilarityMatrix} is used for
feature grouping. Please refer to \code{?MsFeatures::groupSimilarityMatrix}
for details.
}
\references{
Johannes Rainer (2022). MsFeatures: Functionality for Mass Spectrometry
Features. R package version 1.3.0.
'https://github.com/RforMassSpectrometry/MsFeatures

Vincent D. Blondel, Jean-Loup Guillaume, Renaud Lambiotte, Etienne Lefebvre:
Fast unfolding of communities in large networks. J. Stat. Mech. (2008)
P10008

Csardi G, Nepusz T: The igraph software package for complex network
research, InterJournal, Complex Systems 1695. 2006. https://igraph.org
}
\seealso{
See \link{hclust}, \link{cutree}, \link[MsFeatures:groupClosest]{MsFeatures::groupClosest},
\link[MsFeatures:groupConsecutive]{MsFeatures::groupConsecutive}, \link[MsFeatures:groupSimilarityMatrix]{MsFeatures::groupSimilarityMatrix}, and
\link[igraph:cluster_louvain]{igraph::cluster_louvain} for the underlying functions that do work.
}
