% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faahko.R
\docType{data}
\name{faahko_poplin}
\alias{faahko_poplin}
\title{FAAH knockout LC/MS data poplin}
\format{
An object of class \code{poplin} with 206 rows and 12 columns.
}
\usage{
faahko_poplin
}
\description{
A \linkS4class{poplin} object containing FAAH knockout LC/MS feature data
from the \pkg{faahKO} package (\code{faahko3} data) created by coercing the
\link{faahko_se} object.
}
\examples{
faahko_poplin
}
\keyword{datasets}
