% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faahko_se.R
\docType{data}
\name{faahko_se}
\alias{faahko_se}
\title{FAAH knockout LC/MS data SummarizedExperiment}
\format{
An object of class \code{SummarizedExperiment} with 206 rows and 12 columns.
}
\usage{
faahko_se
}
\description{
A \linkS4class{SummarizedExperiment} object containing FAAH knockout LC/MS
feature data from the \pkg{faahKO} (\code{faahko3} data) created with the
\code{xcms::quantify} method.
}
\examples{
faahko_se
}
\references{
Colin A. Smith (2021). faahKO: Saghatelian et al. (2004) FAAH knockout LC/MS
data. R package version 1.32.0. http://dx.doi.org/10.1021/bi0480335
}
\keyword{datasets}
