% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation-methods.R
\name{impute_randomforest}
\alias{impute_randomforest}
\alias{impute_randomforest,matrix-method}
\alias{impute_randomforest,poplin-method}
\title{Random forest imputation}
\usage{
\S4method{impute_randomforest}{matrix}(x, ...)

\S4method{impute_randomforest}{poplin}(x, xin, xout, ...)
}
\arguments{
\item{x}{a matrix or \linkS4class{poplin} object.}

\item{...}{additional arguments passed to \link[missForest]{missForest}.}

\item{xin}{character specifying the name of data to retrieve from \code{x}
when \code{x} is a poplin object.}

\item{xout}{character specifying the name of data to store in \code{x} when
\code{x} is a poplin object.}
}
\value{
a matrix or \linkS4class{poplin} object of the same dimension as
\code{x} containing the imputed intensities.
}
\description{
Apply random forest imputation to a matrix or \linkS4class{poplin} object.
This is an interface to the \link[missForest]{missForest} function from the
\pkg{missForest} package. Since the random forest is a tree-based method, it
is invariant to monotonic transformations.
}
\examples{

if (requireNamespace("missForest", quietly = TRUE)) {
  ## poplin object
  impute_randomforest(faahko_poplin, xin = "raw", xout = "rf")

  ## matrix
  m <- poplin_raw(faahko_poplin, xin = "raw")
  impute_randomforest(m)
}
}
\references{
Daniel J. Stekhoven (2013). missForest: Nonparametric Missing Value
Imputation using Random Forest. R package version 1.4.

Stekhoven D. J., & Buehlmann, P. (2012). MissForest - non-parametric missing
value imputation for mixed-type data. Bioinformatics, 28(1), 112-118.
}
\seealso{
Other imputation methods: 
\code{\link{impute_knn}()},
\code{\link{impute_pca}()},
\code{\link{impute_univariate}()},
\code{\link{poplin_impute}()}
}
\concept{imputation methods}
