% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation-methods.R
\name{impute_simple}
\alias{impute_simple}
\alias{impute_simple,poplin-method}
\title{Simple univariate imputation}
\usage{
\S4method{impute_simple}{matrix}(x, type = c("halfmin", "median", "mean"))

\S4method{impute_simple}{poplin}(x, poplin_in, poplin_out, type = c("halfmin", "median", "mean"))
}
\arguments{
\item{x}{A matrix or \linkS4class{poplin} object.}

\item{type}{A method for doing univariate imputation.}

\item{poplin_in}{Name of a data matrix to retrieve.}

\item{poplin_out}{Name of a data matrix to store.}
}
\value{
A matrix or \linkS4class{poplin} object of the same dimension as
\code{x} containing the imputed intensities.
}
\description{
Apply univariate imputation to a matrix or \linkS4class{poplin} object. The
supported methods include
\itemize{
\item Half-minimum imputation: for each feature, missing values are replaced
with half the observed minimum.
\item Median imputation: for each feature, missing values are replaced with
the median of non-missing values.
\item Mean imputation: for each feature, missing values are replaced with
the mean of non-missing values.
}
}
\references{
Wei, R., Wang, J., Su, M. et al. Missing Value Imputation Approach for Mass
Spectrometry-based Metabolomics Data. Sci Rep 8, 663 (2018).
https://doi.org/10.1038/s41598-017-19120-0
}
\seealso{
Other imputation methods: 
\code{\link{impute_knn}()},
\code{\link{impute_pca}()},
\code{\link{impute_randomforest}()},
\code{\link{poplin_impute}()}
}
\concept{imputation methods}
