% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation-methods.R
\name{impute_univariate}
\alias{impute_univariate}
\alias{impute_univariate,matrix-method}
\alias{impute_univariate,poplin-method}
\title{Univariate imputation}
\usage{
\S4method{impute_univariate}{matrix}(x, type = c("halfmin", "median", "mean"))

\S4method{impute_univariate}{poplin}(x, xin, xout, type = c("halfmin", "median", "mean"))
}
\arguments{
\item{x}{a matrix or \linkS4class{poplin} object.}

\item{type}{character specifying the method for univariate imputation.}

\item{xin}{character specifying the name of data to retrieve from \code{x}
when \code{x} is a poplin object.}

\item{xout}{character specifying the name of data to store in \code{x} when
\code{x} is a poplin object.}
}
\value{
a matrix or \linkS4class{poplin} object of the same dimension as
\code{x} containing the imputed intensities.
}
\description{
Apply univariate imputation to a matrix or \linkS4class{poplin} object. The
supported methods include
\itemize{
\item Half-minimum imputation: for each feature, missing values are replaced
with half the observed minimum.
\item Median imputation: for each feature, missing values are replaced with
the median of non-missing values.
\item Mean imputation: for each feature, missing values are replaced with
the mean of non-missing values.
}
}
\examples{

## poplin object
impute_univariate(faahko_poplin, xin = "raw", xout = "halfmin",
                  type = "halfmin")
## matrix
m <- poplin_raw(faahko_poplin, "raw")
impute_univariate(m, type = "median")
}
\references{
Wei, R., Wang, J., Su, M. et al. Missing Value Imputation Approach for Mass
Spectrometry-based Metabolomics Data. Sci Rep 8, 663 (2018).
https://doi.org/10.1038/s41598-017-19120-0
}
\seealso{
Other imputation methods: 
\code{\link{impute_knn}()},
\code{\link{impute_pca}()},
\code{\link{impute_randomforest}()},
\code{\link{poplin_impute}()}
}
\concept{imputation methods}
