% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization-methods.R
\name{normalize_cyclicloess}
\alias{normalize_cyclicloess}
\alias{normalize_cyclicloess,matrix-method}
\alias{normalize_cyclicloess,poplin-method}
\title{Cyclic LOESS normalization}
\usage{
\S4method{normalize_cyclicloess}{matrix}(
  x,
  pre_log2 = TRUE,
  type = c("fast", "affy", "pairs"),
  span = 0.7,
  iterations = 3
)

\S4method{normalize_cyclicloess}{poplin}(
  x,
  xin,
  xout,
  pre_log2 = TRUE,
  type = c("fast", "affy", "pairs"),
  span = 0.7,
  iterations = 3
)
}
\arguments{
\item{x}{a matrix or \linkS4class{poplin} object.}

\item{pre_log2}{logical controlling whether feature intensities are
log2-transformed before normalization.}

\item{type}{character specifying which variant of the cyclic LOESS
method to use.}

\item{span}{span of LOESS smoothing window, between 0 and 1.}

\item{iterations}{number of times to cycle through all pairs of columns.}

\item{xin}{character specifying the name of data to retrieve from \code{x}
when \code{x} is a poplin object.}

\item{xout}{character specifying the name of data to store in \code{x} when
\code{x} is a poplin object.}
}
\value{
a matrix or \linkS4class{poplin} object of the same dimension as
\code{x} containing the normalized intensities.
}
\description{
Apply Cyclic LOESS normalization to a matrix or \linkS4class{poplin} object.
This is an interface to the \link[limma:normalizeCyclicLoess]{limma::normalizeCyclicLoess} from the
\pkg{limma} package. The input \code{x} is expected to contain
log-transformed raw intensities. See Bolstad et al. (2003) and Ballman et
al. (2004) for details.
}
\examples{

if (requireNamespace("limma", quietly = TRUE)) {
  ## poplin object
  normalize_cyclicloess(faahko_poplin, xin = "knn", xout = "knn_cyclic",
                       pre_log2 = TRUE)
  ## matrix
  m <- poplin_data(faahko_poplin, "knn")
  normalize_cyclicloess(m, pre_log2 = TRUE)
}
}
\references{
Ritchie ME, Phipson B, Wu D, Hu Y, Law CW, Shi W, Smyth GK. limma powers
differential expression analyses for RNA-sequencing and microarray studies.
Nucleic Acids Res. 2015 Apr 20;43(7):e47. doi: 10.1093/nar/gkv007. Epub 2015
Jan 20. PMID: 25605792; PMCID: PMC4402510.

Bolstad BM, Irizarry RA, Astrand M, Speed TP. A comparison of normalization
methods for high density oligonucleotide array data based on variance and
bias. Bioinformatics. 2003 Jan 22;19(2):185-93. doi:
10.1093/bioinformatics/19.2.185. PMID: 12538238.

Ballman KV, Grill DE, Oberg AL, Therneau TM. Faster cyclic loess:
normalizing RNA arrays via linear models. Bioinformatics. 2004 Nov
1;20(16):2778-86. doi: 10.1093/bioinformatics/bth327. Epub 2004 May 27.
PMID: 15166021.
}
\seealso{
Other normalization methods: 
\code{\link{normalize_mad}()},
\code{\link{normalize_mean}()},
\code{\link{normalize_median}()},
\code{\link{normalize_pqn}()},
\code{\link{normalize_scale}()},
\code{\link{normalize_sum}()},
\code{\link{normalize_vsn}()},
\code{\link{poplin_normalize}()}
}
\concept{normalization methods}
