% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization-methods.R
\name{normalize_mad}
\alias{normalize_mad}
\alias{normalize_mad,matrix-method}
\alias{normalize_mad,poplin-method}
\title{Median absolute deviation (MAD) normalization}
\usage{
\S4method{normalize_mad}{matrix}(x, restrict = FALSE, rescale = FALSE)

\S4method{normalize_mad}{poplin}(x, xin, xout, restrict = FALSE, rescale = FALSE)
}
\arguments{
\item{x}{A matrix or \linkS4class{poplin} object.}

\item{restrict}{Logical controlling whether any feature with missing values
is excluded from the calculation of normalization factors.}

\item{rescale}{Logical controlling whether the normalized intensities are
multiplied by the median of normalization factors to make look similar to
their original scales.}

\item{xin}{Character specifying the name of data to retrieve from \code{x}
when \code{x} is a poplin object.}

\item{xout}{Character specifying the name of data to store in \code{x} when
\code{x} is a poplin object.}
}
\value{
A matrix or \linkS4class{poplin} object of the same dimension as
\code{x} containing the normalized intensities.
}
\description{
Apply median absolute deviation (MAD) normalization to a matrix or
\linkS4class{poplin} object. For each sample, feature intensities are
divided by its MAD.
}
\examples{

data(faahko_poplin)

## poplin object
normalize_mad(faahko_poplin, xin = "knn", xout = "knn_mad")

## matrix
m <- poplin_data(faahko_poplin, "knn")
normalize_mad(m, rescale = TRUE)
}
\seealso{
Other normalization methods: 
\code{\link{normalize_cyclicloess}()},
\code{\link{normalize_mean}()},
\code{\link{normalize_median}()},
\code{\link{normalize_pqn}()},
\code{\link{normalize_scale}()},
\code{\link{normalize_sum}()},
\code{\link{normalize_vsn}()},
\code{\link{poplin_normalize}()}
}
\concept{normalization methods}
