% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization-methods.R
\name{normalize_median}
\alias{normalize_median}
\alias{normalize_median,matrix-method}
\alias{normalize_median,poplin-method}
\title{Median normalization}
\usage{
\S4method{normalize_median}{matrix}(x, restrict = FALSE, rescale = FALSE)

\S4method{normalize_median}{poplin}(x, xin, xout, restrict = FALSE, rescale = FALSE)
}
\arguments{
\item{x}{a matrix or \linkS4class{poplin} object.}

\item{restrict}{logical controlling whether any feature with missing values
is excluded from the calculation of normalization factors.}

\item{rescale}{logical controlling whether the normalized intensities are
multiplied by the median of normalization factors to make look similar to
their original scales.}

\item{xin}{character specifying the name of data to retrieve from \code{x}
when \code{x} is a poplin object.}

\item{xout}{character specifying the name of data to store in \code{x} when
\code{x} is a poplin object.}
}
\value{
a matrix or \linkS4class{poplin} object of the same dimension as
\code{x} containing the normalized intensities.
}
\description{
Apply median normalization to a matrix or \linkS4class{poplin} object. For
each sample, feature intensities are divided by its median. The median of
intensity values for individual samples will be one as a result.
}
\examples{
## poplin object
normalize_median(faahko_poplin, xin = "knn", xout = "knn_med")

## matrix
m <- poplin_data(faahko_poplin, "knn")
normalize_median(m, rescale = TRUE)
}
\seealso{
Other normalization methods: 
\code{\link{normalize_cyclicloess}()},
\code{\link{normalize_mad}()},
\code{\link{normalize_mean}()},
\code{\link{normalize_pqn}()},
\code{\link{normalize_scale}()},
\code{\link{normalize_sum}()},
\code{\link{normalize_vsn}()},
\code{\link{poplin_normalize}()}
}
\concept{normalization methods}
