% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization-methods.R
\name{normalize_pqn}
\alias{normalize_pqn}
\alias{normalize_pqn,matrix-method}
\alias{normalize_pqn,poplin-method}
\title{Probabilistic quotient normalization (PQN)}
\usage{
\S4method{normalize_pqn}{matrix}(
  x,
  ref_samples = colnames(x),
  min_frac = 0.5,
  type = c("mean", "median")
)

\S4method{normalize_pqn}{poplin}(
  x,
  xin,
  xout,
  ref_samples = colnames(x),
  min_frac = 0.5,
  type = c("mean", "median")
)
}
\arguments{
\item{x}{a matrix or \linkS4class{poplin} object.}

\item{ref_samples}{a vector of sample names or indices to specify reference
samples for the calculation of quotients. Must be a subset of
\code{colnames(x)} if it is a character vector. If \code{NULL}, all
samples are used.}

\item{min_frac}{a minimum proportion of reference samples for features to be
considered in the calculation of a reference spectrum, between 0 and 1.
For \code{min_frac = 1}, all reference samples should have non-missing
values for the features to be included.}

\item{type}{a method to compute a reference spectrum. Either "mean" or
"median".}

\item{xin}{character specifying the name of data to retrieve from \code{x}
when \code{x} is a poplin object.}

\item{xout}{character specifying the name of data to store in \code{x} when
\code{x} is a poplin object.}
}
\value{
a matrix or \linkS4class{poplin} object of the same dimension as
\code{x} containing the normalized intensities.
}
\description{
Apply probabilistic quotient normalization (PQN) to a matrix or
\linkS4class{poplin} object. For the calculation of quotients, a reference
spectrum needs to be obtained from a mean or median spectrum based on all
spectra of the study or a subset of the study. Feature intensities are
normalized by the median of quotients. See Dieterle et al. (2006) for
details.
}
\examples{

## poplin object
normalize_pqn(faahko_poplin, xin = "knn", xout = "knn_pqn")

## matrix
m <- poplin_data(faahko_poplin, "knn")
normalize_pqn(m)
}
\references{
Dieterle F, Ross A, Schlotterbeck G, Senn H. Probabilistic quotient
normalization as robust method to account for dilution of complex biological
mixtures. Application in 1H NMR metabonomics. Anal Chem. 2006 Jul
1;78(13):4281-90. doi: 10.1021/ac051632c. PMID: 16808434.
}
\seealso{
Other normalization methods: 
\code{\link{normalize_cyclicloess}()},
\code{\link{normalize_mad}()},
\code{\link{normalize_mean}()},
\code{\link{normalize_median}()},
\code{\link{normalize_scale}()},
\code{\link{normalize_sum}()},
\code{\link{normalize_vsn}()},
\code{\link{poplin_normalize}()}
}
\concept{normalization methods}
