% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization-methods.R
\name{normalize_scale}
\alias{normalize_scale}
\alias{normalize_scale,matrix-method}
\alias{normalize_scale,poplin-method}
\title{Feature-based scaling}
\usage{
\S4method{normalize_scale}{matrix}(x, type = c("auto", "range", "pareto", "vast", "level"))

\S4method{normalize_scale}{poplin}(
  x,
  xin,
  xout,
  type = c("auto", "range", "pareto", "vast", "level")
)
}
\arguments{
\item{x}{a matrix or \linkS4class{poplin} object.}

\item{type}{the scaling method to be applied.}

\item{xin}{character specifying the name of data to retrieve from \code{x}
when \code{x} is a poplin object.}

\item{xout}{character specifying the name of data to store in \code{x} when
\code{x} is a poplin object.}
}
\value{
a matrix or \linkS4class{poplin} object of the same dimension as
\code{x} containing the normalized intensities.
}
\description{
Apply feature-based scaling to a matrix or \linkS4class{poplin} object. The
supported methods include
\itemize{
\item Auto scaling (unit variance scaling): each feature is scaled by its
standard deviation.
\item Range scaling: each feature is scaled by its range.
\item Pareto scaling: each feature is scaled by the square root of its standard
deviation.
\item Vast scaling (variance stability scaling): it is an extension of auto
scaling, using the product of standard deviation and coefficient of
variation as a scaling factor.
\item Level scaling: each feature is scaled by its mean.
}
Note that each feature is mean-centered prior to division. See van den Berg
et al. (2006) for details.
}
\examples{

## poplin object
normalize_scale(faahko_poplin, xin = "knn", xout = "knn_auto", type = "auto")

## matrix
m <- poplin_data(faahko_poplin, "knn")
normalize_scale(m, "pareto")
}
\references{
van den Berg RA, Hoefsloot HC, Westerhuis JA, Smilde AK, van der Werf MJ.
Centering, scaling, and transformations: improving the biological
information content of metabolomics data. BMC Genomics. 2006 Jun 8;7:142.
doi: 10.1186/1471-2164-7-142. PMID: 16762068; PMCID: PMC1534033.
}
\seealso{
Other normalization methods: 
\code{\link{normalize_cyclicloess}()},
\code{\link{normalize_mad}()},
\code{\link{normalize_mean}()},
\code{\link{normalize_median}()},
\code{\link{normalize_pqn}()},
\code{\link{normalize_sum}()},
\code{\link{normalize_vsn}()},
\code{\link{poplin_normalize}()}
}
\concept{normalization methods}
