% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBox.R
\name{plotBox}
\alias{plotBox}
\title{Box plot}
\usage{
plotBox(x, xin, group, log2 = FALSE, violin = FALSE, ylab = "Intensity")
}
\arguments{
\item{x}{A matrix-like object or \linkS4class{SummarizedExperiment} object.}

\item{xin}{A string specifying which assay values to use when \code{x} is a
SummarizedExperiment object.}

\item{group}{A discrete variable to visualize the grouping structure.}

\item{log2}{A logical specifying whether feature intensities needs to be
log2-transformed before visualization.}

\item{violin}{A logical specifying whether a violin plot is shown instead of
a boxplot.}

\item{ylab}{A string specifying the title of y-axis.}
}
\value{
A ggplot object.
}
\description{
Produces a box-and-whisker plot with a \linkS4class{SummarizedExperiment} or
matrix-like object where rows represent features and columns represent
samples.
}
\examples{

data(faahko_poplin)

## sample group variable
group <- colData(faahko_poplin)$sample_group

## poplin object
poplin_boxplot(faahko_poplin, xin = "knn_cyclic", group = group)

## matrix
m <- poplin_data(faahko_poplin, "knn_cyclic")
poplin_boxplot(m, group = group)
}
