% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poplin.R
\name{poplin-class}
\alias{poplin-class}
\alias{poplin}
\alias{coerce,SummarizedExperiment,poplin-method}
\title{The poplin class}
\usage{
## Constructor
poplin(..., poplin_data_list = list(), poplin_reduced_list = list())
}
\arguments{
\item{...}{arguments passed to the \link{SummarizedExperiment} constructor.}

\item{poplin_data_list}{a list of matrix-like objects containing data
processing results.}

\item{poplin_reduced_list}{a list of matrix-like objects containing dimension
reduction results.}
}
\value{
a poplin object
}
\description{
The poplin class is designed to process LC/MS data. It is an extension of
the standard SummarizedExperiment class, and supports additional containers
for data processing results (e.g., normalization, imputation) via
\link{poplin_data_list} and dimension reduction results (PCA, PLS-DA) via
\link{poplin_reduced_list}.
}
\examples{
nsamp <- 10
nfeature <- 200
intensity <- rlnorm(nsamp * nfeature, 10, 1)
m <- matrix(intensity, nrow = nfeature, ncol = nsamp)
rownames(m) <- paste0("F", 1:nrow(m))
colnames(m) <- paste0("S", 1:ncol(m))
poplin(assays = list(raw = m))

## Coercion from an SummarizedExperiment object
se <- SummarizedExperiment(assays = list(raw = m))
as(se, "poplin")

}
\author{
Jaehyun Joo
}
