% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine-methods.R
\docType{methods}
\name{poplin-combine/subset}
\alias{poplin-combine/subset}
\title{Combining or subsetting poplin objects}
\description{
Methods to combine multiple \linkS4class{poplin} objects by row or column,
or to subset a \linkS4class{poplin} object by row or column.
}
\section{Combining}{


In the code snippets below, \code{...} represents multiple
\linkS4class{poplin} objects to be combined.

\describe{
\item{\code{rbind(..., deparse.level = 1)}:}{

Take a sequence of poplin objects in \code{...} and combine by rows. Note
that all objects in ... must have the exact same values for
\link{poplin_data_list}. See \code{?}\link{rbind} for the interpretation of
\code{deparse.level}.

}
\item{\code{cbind(..., deparse.level = 1)}:}{

Take a sequence of poplin objects in \code{...} and combine by columns. Note
that all objects in ... must have the same values of \link{poplin_reduced_names}.
Dimension reduction results with the same name across the objects will be
combined row-wise to create the corresponding entry in the output object.
See \code{?}\link{cbind} for the interpretation of \code{deparse.level}.

}
}

Refer to \code{help("SummarizedExperiment-class", package =
"SummarizedExperiment")} for details on how \code{rowData}, \code{colData},
and \code{metadata} are combined in the output object.
}

\section{Subsetting}{


In the code snippets below, \code{x} is a \linkS4class{poplin}
object.

\describe{

\item{\code{x[i, j]}, \code{x[i, j] <- value}:}{\code{i} and \code{j} are
indices specifying the rows and columns to extract or replace. Indices can
be a logical, integer, character vector, or empty (missing). \code{value}
must be a poplin object with the dimension and assay elements consistent
with the subset \code{x[i, j]} being replaced. \code{value} is also expected
to have the same name and order of \link{poplin_data_names} (and
\link{poplin_reduced_names} as well if \code{j} is specified) as \code{x}.}

}
}

