% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poplin.R
\name{poplin-internals}
\alias{poplin-internals}
\alias{poplinData}
\alias{poplinData<-}
\alias{poplinReduced}
\alias{poplinReduced<-}
\title{Internal poplin fields}
\description{
Methods to get or set internal fields in a \linkS4class{poplin} object.
These methods are not intended to be used by end users of the \pkg{poplin}
package.
}
\details{
In the snippets below, \code{x} is a \linkS4class{poplin} object.

\describe{

\item{\code{poplinData(x)}, \code{poplinData(x) <- value}:}{ Returns a
\linkS4class{DataFrame} of matrices containing one or more data processing
results. \code{value} must be a \linkS4class{DataFrame} with the dimension
equal to \code{dim(x)}. End users are supposed to interact with this field
via \link{poplin_data}. }

\item{\code{poplinReduced(x)}, \code{poplinReduced(x) <- value}:}{
Returns a \linkS4class{DataFrame} of matrices containing one or more
dimension-reduced data. \code{value} must be a \linkS4class{DataFrame} with
the dimension equal to \code{ncol(x)}. End users are supposed to interact
with this field via \link{poplin_reduced}. }

}
}
\seealso{
\link{poplin_data}, \link{poplin_reduced}
}
