% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poplin.matrix.R
\name{poplin.matrix}
\alias{poplin.matrix}
\alias{poplin.matrix-class}
\alias{[.poplin.matrix}
\alias{rbind.poplin.matrix}
\alias{cbind.poplin.matrix}
\title{The poplin.matrix class}
\description{
A matrix class that retains its attributes upon being subsetted or combined.
It is a simply renamed version of \code{reduced.dim.matrix} in the
\pkg{SingleCellExperiment}. This is useful for storing metadata about a
dimensionality reduction result alongside the matrix, and for ensuring that
the metadata persists when the matrix is stored inside
\link{poplin_reduced_list}.
}
\references{
Amezquita, R.A., Lun, A.T.L., Becht, E. et al. Orchestrating single-cell
analysis with Bioconductor. Nat Methods 17, 137–145 (2020).
https://doi.org/10.1038/s41592-019-0654-x
}
\keyword{internal}
