% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poplin.R
\name{poplin}
\alias{poplin}
\title{The poplin class}
\usage{
poplin(intensity, ...)
}
\arguments{
\item{intensity}{Peak intensity matrix where rows represent features and
columns represent samples.}

\item{...}{Arguments passed to the \link{SummarizedExperiment} constructor.}
}
\value{
A poplin object
}
\description{
The poplin class is designed to process LC/MS data. It is an extension of
the standard SummarizedExperiment class, and supports additional containers
for data processing results (e.g., normalization, imputation) via
\link{poplinData} and dimension reduction results (PCA, PLS-DA) via
\link{poplinReducedData}. A poplin object also can be created by coercing from a
\linkS4class{SummarizedExperiment} object.
}
\examples{
nsamp <- 20
nfeature <- 1000
m <- matrix(sample(1:100000, size = 200000), nrow = nsamp, ncol = nfeature)
poplin(m)
## Coercion from an SummarizedExperiment object
as(se, "poplin")

}
\author{
Jaehyun Joo
}
