% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-biplot.R
\name{poplin_biplot}
\alias{poplin_biplot}
\alias{poplin_biplot.default}
\alias{poplin_biplot.poplin.pca}
\alias{poplin_biplot.poplin.plsda}
\alias{poplin_biplot.poplin}
\title{Biplot of dimension-reduced data}
\usage{
\method{poplin_biplot}{default}(
  x,
  y,
  comp = 1:2,
  group,
  group_col = NULL,
  point_size = 1.5,
  point_shape_by_group = FALSE,
  label = FALSE,
  label_size = 3.88,
  label_subset = NULL,
  ellipse = FALSE,
  xlab = NULL,
  ylab = NULL,
  title = NULL,
  legend = TRUE,
  arrow_len = 0.2,
  arrow_col = "red",
  arrow_alpha = 0.3,
  arrow_label = TRUE,
  arrow_label_ext = 1.1,
  arrow_label_size = 3.88,
  arrow_label_col = "red",
  arrow_label_subset = NULL
)

\method{poplin_biplot}{poplin.pca}(x, scale = 1, comp = 1:2, group, ...)

\method{poplin_biplot}{poplin.plsda}(x, comp = 1:2, group = attr(x, "Y.observed"), ...)

\method{poplin_biplot}{poplin}(x, poplin_in, comp = 1:2, ...)
}
\arguments{
\item{x}{A dimension-reduced data matrix produced by \link{poplin_reduce} or
\linkS4class{poplin} object containing dimension-reduced data.}

\item{y}{A data matrix for loadings. Not required for the \link{reduce_pca} and
\link{reduce_plsda} outputs.}

\item{comp}{A numeric vector of length 2 indicating two components to plot.}

\item{group}{A discrete variable to visualize the grouping structure.}

\item{group_col}{A vector of colors with the same length of unique values in
\code{group}.}

\item{point_size}{The size of points.}

\item{point_shape_by_group}{Logical controlling whether each group have
different shapes of data points. Also can be a numeric vector with the
same length of unique values in \code{group} to manually set point shapes.}

\item{label}{Logical controlling whether score labels are shown instead of
points.}

\item{label_subset}{A character vector specifying a subset of score labels
to display.}

\item{ellipse}{Logical controlling whether data ellipses are shown using
\link[ggplot2:stat_ellipse]{ggplot2::stat_ellipse}.}

\item{xlab}{The title of x-axis of the plot.}

\item{ylab}{The title of y-axis of the plot.}

\item{title}{The main title of the plot.}

\item{legend}{Logical controlling whether the plot legend is shown.}

\item{arrow_len}{Numeric value controlling the length of arrow head.}

\item{arrow_alpha}{Numeric value controlling the transparency of arrow.}

\item{arrow_label}{Logical controlling whether text labels for arrows are
shown.}

\item{arrow_label_ext}{Numeric value controlling the scalar extension for
arrow labels.}

\item{arrow_label_size}{Numeric value controlling the size of arrow labels.}

\item{arrow_label_col}{Character string indicating the color of arrow
labels.}

\item{arrow_label_subset}{A character vector specifying a subset of arrow
labels to display.}

\item{...}{Arguments passed to the default method.}

\item{poplin_in}{Name of a dimension-reduced data matrix to retrieve.}
}
\value{
A ggplot object.
}
\description{
Visualize an overlay of a score plot and a loading plot using the
\link{poplin_reduce} output.
}
\seealso{
\link{poplin_reduce}, \link{poplin_scoreplot}.
}
