% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-boxplot.R
\name{poplin_boxplot}
\alias{poplin_boxplot}
\alias{poplin_boxplot.default}
\alias{poplin_boxplot.poplin}
\title{Box plot}
\usage{
\method{poplin_boxplot}{default}(
  x,
  group,
  pre_log2 = FALSE,
  violin = FALSE,
  ylab = "Intensity",
  ...
)

\method{poplin_boxplot}{poplin}(
  x,
  xin,
  group,
  pre_log2 = FALSE,
  violin = FALSE,
  ylab = "Intensity",
  ...
)
}
\arguments{
\item{x}{a matrix or \linkS4class{poplin} object.}

\item{group}{a discrete variable to visualize the grouping structure.}

\item{pre_log2}{logical controlling whether feature intensities are
log2-transformed before visualization.}

\item{violin}{logical controlling whether a violin plot is drawn instead of
the boxplot.}

\item{ylab}{the title of y-axis of the plot.}

\item{...}{reserved for future use.}

\item{xin}{character specifying the name of data to retrieve from \code{x}
when \code{x} is a poplin object}
}
\value{
a ggplot object.
}
\description{
Produce a box-and-whisker plot of the feature intensity values.
}
\examples{

## sample group variable
group <- colData(faahko_poplin)$sample_group

## poplin object
poplin_boxplot(faahko_poplin, xin = "knn_cyclic", group = group)

## matrix
m <- poplin_data(faahko_poplin, "knn_cyclic")
poplin_boxplot(m, group = group)
}
