% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-corplot.R
\name{poplin_corplot}
\alias{poplin_corplot}
\alias{poplin_corplot.default}
\alias{poplin_corplot.poplin}
\title{Correlation plot}
\usage{
\method{poplin_corplot}{default}(
  x,
  unit = c("sample", "feature"),
  use = c("everything", "all.obs", "complete.obs", "na.or.complete",
    "pairwise.complete.obs"),
  method = c("pearson", "kendall", "spearman"),
  widths = NULL,
  heights = NULL,
  colors = viridis::viridis(n = 256, alpha = 1, begin = 0, end = 1, option = "viridis"),
  label = FALSE,
  digits = 2,
  grid_gap = 0,
  hide_colorbar = FALSE,
  showticklabels = c(TRUE, TRUE),
  row_dend_left = FALSE,
  ...
)

\method{poplin_corplot}{poplin}(
  x,
  poplin_in,
  unit = c("sample", "feature"),
  use = c("everything", "all.obs", "complete.obs", "na.or.complete",
    "pairwise.complete.obs"),
  method = c("pearson", "kendall", "spearman"),
  widths = NULL,
  heights = NULL,
  colors = viridis::viridis(n = 256, alpha = 1, begin = 0, end = 1, option = "viridis"),
  grid_gap = 0,
  hide_colorbar = FALSE,
  showticklabels = c(TRUE, TRUE),
  row_dend_left = FALSE,
  ...
)
}
\arguments{
\item{x}{A matrix or \linkS4class{poplin} object.}

\item{unit}{A character string specifying whether a correlation matrix is
computed based on samples or features. One of "sample" or "feature".}

\item{use}{A method for computing correlations in the presence of missing
values. Refer to \code{?cor} for details.}

\item{method}{A character string specifying which correlation coefficient is
to be computed. Refer to \code{?cor} for details.}

\item{widths}{Relative widths of heatmap and dendrogram.}

\item{heights}{Relative heights of heatmap and dendrogram.}

\item{colors}{A vector of colors for heatmap.}

\item{label}{If \code{TRUE}, cell values are shown.}

\item{digits}{The desired number of digits when \code{label = TRUE}.}

\item{grid_gap}{Gap between cells.}

\item{hide_colorbar}{Logical controlling whether the color bar (legend) is
hidden.}

\item{showticklabels}{A logical vector of length 2 (x-axis, y-axis). If
\code{FALSE}, the ticks are removed from the sides of the plot.}

\item{row_dend_left}{Logical controlling whether the row dendrogram is
placed on the left on the plot.}

\item{...}{Additional arguments passed to \link[heatmaply:heatmaply]{heatmaply::heatmaply}.}

\item{poplin_in}{Name of a data matrix to retrieve.}
}
\value{
\code{gtable} of aligned plots.
}
\description{
Visualize correlations between samples or features. All values in a
correlation matrix are visualized by rectangles.
}
