% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation-methods.R
\name{poplin_impute}
\alias{poplin_impute}
\alias{poplin_impute,poplin-method}
\alias{poplin_impute_knn,matrix-method}
\alias{poplin_impute_knn,poplin-method}
\alias{poplin_impute_halfmin,matrix-method}
\alias{poplin_impute_halfmin,poplin-method}
\alias{poplin_impute_median,matrix-method}
\alias{poplin_impute_median,poplin-method}
\alias{poplin_impute_mean,matrix-method}
\alias{poplin_impute_mean,poplin-method}
\alias{poplin_impute_pca,matrix-method}
\alias{poplin_impute_pca,poplin-method}
\alias{poplin_impute_randomforest,matrix-method}
\alias{poplin_impute_randomforest,poplin-method}
\title{Imputation methods}
\usage{
\S4method{poplin_impute}{matrix}(x, method, ...)

\S4method{poplin_impute}{poplin}(x, method, poplin_in, poplin_out, ...)

\S4method{poplin_impute_knn}{matrix}(x, ...)

\S4method{poplin_impute_knn}{poplin}(x, poplin_in, poplin_out, ...)

\S4method{poplin_impute_halfmin}{matrix}(x, ...)

\S4method{poplin_impute_halfmin}{poplin}(x, poplin_in, poplin_out, ...)

\S4method{poplin_impute_median}{matrix}(x, ...)

\S4method{poplin_impute_median}{poplin}(x, poplin_in, poplin_out, ...)

\S4method{poplin_impute_mean}{matrix}(x, ...)

\S4method{poplin_impute_mean}{poplin}(x, poplin_in, poplin_out, ...)

\S4method{poplin_impute_pca}{matrix}(x, ...)

\S4method{poplin_impute_pca}{poplin}(x, poplin_in, poplin_out, ...)

\S4method{poplin_impute_randomforest}{matrix}(x, ...)

\S4method{poplin_impute_randomforest}{poplin}(x, poplin_in, poplin_out, ...)
}
\description{
Missing values are frequently found in metabolomics data. The \pkg{poplin}
package provides a few options to handle them.
}
