% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization-methods.R
\name{poplin_normalize}
\alias{poplin_normalize}
\alias{poplin_normalize,matrix-method}
\alias{poplin_normalize,poplin-method}
\title{Normalization methods}
\usage{
\S4method{poplin_normalize}{matrix}(
  x,
  method = c("pqn", "sum", "mean", "median", "mad", "cyclicloess", "vsn", "scale"),
  ...
)

\S4method{poplin_normalize}{poplin}(
  x,
  method = c("pqn", "sum", "mean", "median", "mad", "cyclicloess", "vsn", "scale"),
  xin,
  xout,
  ...
)
}
\arguments{
\item{x}{A matrix or \linkS4class{poplin} object.}

\item{method}{The normalization method to be used, defaulting to "pqn".}

\item{...}{Arguments passed to a specific normalization method.}

\item{xin}{Character specifying the name of data to retrieve from \code{x}
when \code{x} is a poplin object.}

\item{xout}{Character specifying the name of data to store in \code{x} when
\code{x} is a poplin object.}
}
\value{
A matrix or \linkS4class{poplin} object of the same dimension as
\code{x} containing the normalized intensities.
}
\description{
Data normalization aims to reduce unwanted variations. The \pkg{poplin}
package provides a few data-driven normalization methods.
\link{poplin_normalize} is a wrapper for the following set of functions:
\describe{
\item{\code{\link{normalize_pqn}}:}{
PQN (probabilistic quotient normalization)
}
\item{\code{\link{normalize_sum}}:}{
sum normalization
}
\item{\code{\link{normalize_mean}}:}{
mean normalization
}
\item{\code{\link{normalize_median}}:}{
median normalization
}
\item{\code{\link{normalize_mad}}:}{
MAD (median absolute deviation) normalization
}
\item{\code{\link{normalize_cyclicloess}}:}{
cyclic LOESS normalization
}
\item{\code{\link{normalize_vsn}}:}{
VSN (variance stabilizing normalization)
}
\item{\code{\link{normalize_scale}}:}{
feature-based scaling (e.g., auto, range, pareto, vast, level)
}
}
}
\examples{

data(faahko_poplin)

## poplin object
poplin_normalize(faahko_poplin, method = "pqn",
                 xin = "knn", xout = "knn_pqn")

## matrix
m <- poplin_data(faahko_poplin, "knn")
poplin_normalize(m)
}
\seealso{
Other normalization methods: 
\code{\link{normalize_cyclicloess}()},
\code{\link{normalize_mad}()},
\code{\link{normalize_mean}()},
\code{\link{normalize_median}()},
\code{\link{normalize_pqn}()},
\code{\link{normalize_scale}()},
\code{\link{normalize_sum}()},
\code{\link{normalize_vsn}()}
}
\concept{normalization methods}
