% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization-methods.R
\name{poplin_normalize_pqn}
\alias{poplin_normalize_pqn}
\alias{poplin_normalize_pqn,poplin-method}
\title{Probabilistic quotient normalization (PQN)}
\usage{
\S4method{poplin_normalize_pqn}{matrix}(
  x,
  ref_samples = colnames(x),
  min_frac = 0.5,
  type = c("mean", "median")
)

\S4method{poplin_normalize_pqn}{poplin}(
  x,
  poplin_in,
  poplin_out,
  ref_samples = colnames(x),
  min_frac = 0.5,
  type = c("mean", "median")
)
}
\arguments{
\item{x}{A matrix or \linkS4class{poplin} object.}

\item{ref_samples}{A vector of sample names or indices to identify reference
samples for the calculation of quotients. Must be a subset of
\code{colnames(x)} if it is a character vector. If \code{NULL}, all
samples are used.}

\item{min_frac}{A minimum proportion of reference samples for features to
be considered in the calculation of a reference spectrum.}

\item{type}{A method to compute a reference spectrum}

\item{poplin_in}{Name of a data matrix to retrieve.}

\item{poplin_out}{Name of a data matrix to store.}
}
\value{
A matrix or \linkS4class{poplin} object of the same dimension as
\code{x} containing the normalized intensities.
}
\description{
Apply probabilistic quotient normalization (PQN) to a matrix or
\linkS4class{poplin} object. For the calculation of quotients, a reference
spectrum needs to be obtained from a mean or median spectrum based on all
spectra of the study or a subset of the study. Feature intensities are
normalized by the median of quotients. See Dieterle et al. (2006) for
details.
}
\references{
Dieterle F, Ross A, Schlotterbeck G, Senn H. Probabilistic quotient
normalization as robust method to account for dilution of complex biological
mixtures. Application in 1H NMR metabonomics. Anal Chem. 2006 Jul
1;78(13):4281-90. doi: 10.1021/ac051632c. PMID: 16808434.
}
\seealso{
Other normalization methods: 
\code{\link{poplin_normalize_cyclicloess}()},
\code{\link{poplin_normalize_mad}()},
\code{\link{poplin_normalize_mean}()},
\code{\link{poplin_normalize_median}()},
\code{\link{poplin_normalize_scale}()},
\code{\link{poplin_normalize_sum}()},
\code{\link{poplin_normalize_vsn}()},
\code{\link{poplin_normalize}()}
}
\concept{normalization methods}
