% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization-methods.R
\name{poplin_normalize_scale}
\alias{poplin_normalize_scale}
\alias{poplin_normalize_scale,poplin-method}
\title{Feature-based scaling}
\usage{
\S4method{poplin_normalize_scale}{matrix}(x, type = c("auto", "range", "pareto", "vast", "level"))

\S4method{poplin_normalize_scale}{poplin}(
  x,
  poplin_in,
  poplin_out,
  type = c("auto", "range", "pareto", "vast", "level")
)
}
\arguments{
\item{x}{A matrix or \linkS4class{poplin} object.}

\item{type}{A scaling method to be applied.}

\item{poplin_in}{Name of a data matrix to retrieve.}

\item{poplin_out}{Name of a data matrix to store.}
}
\value{
A matrix or \linkS4class{poplin} object of the same dimension as
\code{x} containing the normalized intensities.
}
\description{
Apply feature-based scaling to a matrix or \linkS4class{poplin} object. The
supported methods include
\itemize{
\item Auto scaling (unit variance scaling): each feature is scaled by its
standard deviation.
\item Range scaling: each feature is scaled by its range.
\item Pareto scaling: each feature is scaled by the square root of its standard
deviation.
\item Vast scaling (variance stability scaling): it is an extension of auto
scaling, using the product of standard deviation and coefficient of
variation as a scaling factor.
\item Level scaling: each feature is scaled by its mean.
}
Note that each feature is mean-centered prior to division. See van den Berg
et al. (2006) for details.
}
\references{
van den Berg RA, Hoefsloot HC, Westerhuis JA, Smilde AK, van der Werf MJ.
Centering, scaling, and transformations: improving the biological
information content of metabolomics data. BMC Genomics. 2006 Jun 8;7:142.
doi: 10.1186/1471-2164-7-142. PMID: 16762068; PMCID: PMC1534033.
}
\seealso{
Other normalization methods: 
\code{\link{poplin_normalize_cyclicloess}()},
\code{\link{poplin_normalize_mad}()},
\code{\link{poplin_normalize_mean}()},
\code{\link{poplin_normalize_median}()},
\code{\link{poplin_normalize_pqn}()},
\code{\link{poplin_normalize_sum}()},
\code{\link{poplin_normalize_vsn}()},
\code{\link{poplin_normalize}()}
}
\concept{normalization methods}
