% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization-methods.R
\name{poplin_normalize_sum}
\alias{poplin_normalize_sum}
\alias{poplin_normalize_sum,poplin-method}
\title{Sum normalization (TIC normalization)}
\usage{
\S4method{poplin_normalize_sum}{matrix}(x, restrict = FALSE, rescale = FALSE)

\S4method{poplin_normalize_sum}{poplin}(
  x,
  poplin_in,
  poplin_out,
  restrict = FALSE,
  rescale = FALSE
)
}
\arguments{
\item{x}{A matrix or \linkS4class{poplin} object.}

\item{restrict}{If \code{TRUE}, any feature with missing values is excluded
from the calculation of normalization factors.}

\item{rescale}{If \code{TRUE}, normalized intensities are multiplied by the
median of normalization factors to make look similar to their original
scales.}

\item{poplin_in}{Name of a data matrix to retrieve.}

\item{poplin_out}{Name of a data matrix to store.}
}
\value{
A matrix or \linkS4class{poplin} object of the same dimension as
\code{x} containing the normalized intensities.
}
\description{
Apply sum normalization to a matrix or \linkS4class{poplin} object. For each
sample, feature intensities are divided by its Total Ion Current (TIC),
i.e., every feature is divided by the sum of all intensity values.
}
\seealso{
Other normalization methods: 
\code{\link{poplin_normalize_cyclicloess}()},
\code{\link{poplin_normalize_mad}()},
\code{\link{poplin_normalize_mean}()},
\code{\link{poplin_normalize_median}()},
\code{\link{poplin_normalize_pqn}()},
\code{\link{poplin_normalize_scale}()},
\code{\link{poplin_normalize_vsn}()},
\code{\link{poplin_normalize}()}
}
\concept{normalization methods}
