% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poplinRaw-methods.R
\name{poplin_raw}
\alias{poplin_raw}
\alias{poplin_raw<-}
\alias{poplin_raw_list}
\alias{poplin_raw_list<-}
\alias{poplin_raw_names}
\alias{poplin_raw_names<-}
\title{Raw data methods}
\description{
Methods to get or set raw intensity data. These methods are simply aliases
for \link{assayNames}, \link{assays}, and \link{assay} functions from the
\pkg{SummarizedExperiment} package.
}
\details{
In the code snippets below, \code{x} is a \linkS4class{poplin} object.

\describe{

\item{\code{poplin_raw_names(x), poplin_raw_names(x) <- value}:}{ Get or set
the names of raw intensity data sets. \code{value} must be a character
vector to be assigned. This is an alias of
\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::assayNames}. }

\item{\code{poplin_raw_list(x), poplin_raw_list(x) <- value}:}{ Get or set
the raw intensity data sets. \code{value} must be a named \linkS4class{List}
of matrices in which each element has the same dimension as \code{x}. This
is an alias of \link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::assays}. }

\item{\code{poplin_raw(x, i), poplin_raw(x, i) <- value}:}{ Get or set the
\code{i}-th element of raw intensity data sets. It is a convenient
alternative to \code{poplin_raw_list(x)[[i]]}. \code{value} must be a matrix
with the same dimension as \code{x}. If \code{i} is missing, the first entry
is retrieved or replaced. This is an alias of \link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::assay}.
}

}
}
\seealso{
\link{poplin_data}, \link{poplin_reduced}
}
