% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction-methods.R
\name{poplin_reduce}
\alias{poplin_reduce}
\alias{poplin_reduce,poplin-method}
\alias{poplin_reduce_pca,matrix-method}
\alias{poplin_reduce_pca,poplin-method}
\alias{poplin_reduce_tsne,matrix-method}
\alias{poplin_reduce_tsne,poplin-method}
\alias{poplin_reduce_plsda,matrix-method}
\alias{poplin_reduce_plsda,poplin-method}
\title{Dimension reduction methods}
\usage{
\S4method{poplin_reduce}{matrix}(x, method, ...)

\S4method{poplin_reduce}{poplin}(x, method, poplin_in, poplin_out, ...)

\S4method{poplin_reduce_pca}{matrix}(x, ...)

\S4method{poplin_reduce_pca}{poplin}(x, poplin_in, poplin_out, ...)

\S4method{poplin_reduce_tsne}{matrix}(x, ...)

\S4method{poplin_reduce_tsne}{poplin}(x, poplin_in, poplin_out, ...)

\S4method{poplin_reduce_plsda}{matrix}(x, ...)

\S4method{poplin_reduce_plsda}{poplin}(x, poplin_in, poplin_out, ...)
}
\description{
Dimension reduction methods
}
