% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction-methods.R
\name{reduce_pca}
\alias{reduce_pca}
\alias{reduce_pca,matrix-method}
\alias{reduce_pca,poplin-method}
\title{Principal component analysis (PCA)}
\usage{
\S4method{reduce_pca}{matrix}(x, ncomp = 2, center = TRUE, scale = FALSE, ...)

\S4method{reduce_pca}{poplin}(x, xin, xout, ncomp = 2, center = TRUE, scale = FALSE, ...)
}
\arguments{
\item{x}{A matrix or \linkS4class{poplin} object.}

\item{ncomp}{Output dimensionality.}

\item{center}{Logical indicating mean-centering prior to PCA.}

\item{scale}{Logical indicating unit variance scaling prior to PCA.}

\item{...}{Additional arguments passed to \link[pcaMethods]{nipalsPca}.}

\item{xin}{Character specifying the name of data to retrieve from \code{x}
when \code{x} is a poplin object.}

\item{xout}{character specifying the name of data to store in \code{x} when
\code{x} is a poplin object.}
}
\value{
A poplin.pca or \linkS4class{poplin} object with the same number of
rows as \code{ncol(x)} containing the dimension reduction result.
poplin.pca is a matrix containing custom attributes used to summarize and
visualize the PCA result.
}
\description{
Apply PCA to a matrix or \linkS4class{poplin} object. For the data without
missing values, PCA is performed via singular value decomposition.
Otherwise, PCA is performed using the non-linear iterative partial least
squares (NIPALS) algorithm via the \link[pcaMethods]{nipalsPca} function
from the \pkg{pcaMethods} package. NIPALS PCA can handle a small amount of
missing values.
}
\examples{

data(faahko_poplin)

## poplin object
out <- reduce_pca(faahko_poplin, xin = "knn_cyclic", xout = "pca")
summary(poplin_reduced(out, "pca"))

## matrix
m <- poplin_data(faahko_poplin, "knn_cyclic")
out <- reduce_pca(m, ncomp = 3)
summary(out)
}
\references{
Wold, H. (1966). Estimation of principal components and related models by
iterative least squares. In P. R. Krishnajah (Ed.), Multivariate analysis
(pp. 391-420). NewYork: Academic Press.
}
\seealso{
Other data reduction methods: 
\code{\link{poplin_reduce}()},
\code{\link{reduce_plsda}()},
\code{\link{reduce_tsne}()}
}
\concept{data reduction methods}
