% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction-methods.R
\name{reduce_plsda}
\alias{reduce_plsda}
\alias{reduce_plsda,matrix-method}
\alias{reduce_plsda,poplin-method}
\title{Partial least squares-discriminant analysis (PLS-DA)}
\usage{
\S4method{reduce_plsda}{matrix}(x, y, ncomp = 2, center = TRUE, scale = FALSE, ...)

\S4method{reduce_plsda}{poplin}(x, xin, xout, y, ncomp = 2, center = TRUE, scale = FALSE, ...)
}
\arguments{
\item{x}{A matrix or \linkS4class{poplin} object.}

\item{y}{A factor vector for discrete outcome.}

\item{ncomp}{output dimensionality.}

\item{center}{Logical indicating mean-centering prior to PLS-DA.}

\item{scale}{Logical indicating unit variance scaling prior to PLS-DA.}

\item{...}{Additional arguments passed to \link[pls]{plsr}.}

\item{xin}{character specifying the name of data to retrieve from \code{x}
when \code{x} is a poplin object.}

\item{xout}{character specifying the name of data to store in \code{x} when
\code{x} is a poplin object.}
}
\value{
A poplin.plsda or \linkS4class{poplin} object with the same number of
rows as \code{ncol(x)} containing the dimension reduction result.
poplin.plsda is a matrix containing custom attributes used to summarize and
visualize the PLS-DA result.
}
\description{
Apply PLS-DA to a matrix or \linkS4class{poplin} object. It performs
standard PLS for classification using the \link[pls]{plsr} function from the
\pkg{pls} package.
}
\examples{

data(faahko_poplin)

if (requireNamespace("pls", quietly = TRUE)) {
  ## response vector
  y <- factor(colData(faahko_poplin)$sample_group, levels = c("WT", "KO"))

  ## poplin object
  out <- reduce_plsda(faahko_poplin, xin = "knn_cyclic", xout = "plsda",
                      y = y)
  summary(poplin_reduced(out, "plsda"))

  ## matrix
  m <- poplin_data(faahko_poplin, "knn_cyclic")
  out <- reduce_plsda(m, y = y, ncomp = 3)
  summary(out)
}
}
\references{
Kristian Hovde Liland, Bjørn-Helge Mevik and Ron Wehrens (2021). pls:
Partial Least Squares and Principal Component Regression. R package version
2.8-0. https://CRAN.R-project.org/package=pls
}
\seealso{
Other data reduction methods: 
\code{\link{poplin_reduce}()},
\code{\link{reduce_pca}()},
\code{\link{reduce_tsne}()}
}
\concept{data reduction methods}
