% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction-methods.R
\name{reduce_tsne}
\alias{reduce_tsne}
\alias{reduce_tsne,matrix-method}
\alias{reduce_tsne,poplin-method}
\title{t-distributed stochastic neighbor embedding (t-SNE)}
\usage{
\S4method{reduce_tsne}{matrix}(x, ncomp = 2, normalize = TRUE, ...)

\S4method{reduce_tsne}{poplin}(x, xin, xout, ncomp = 2, normalize = TRUE, ...)
}
\arguments{
\item{x}{A matrix or \linkS4class{poplin} object.}

\item{ncomp}{Output dimensionality.}

\item{normalize}{Logical controlling whether the input matrix is
mean-centered and scaled so that the largest absolute of the centered
matrix is equal to unity. See \link[Rtsne]{normalize_input} for details.}

\item{...}{Additional arguments passed to \link[Rtsne]{Rtsne}.}

\item{xin}{character specifying the name of data to retrieve from \code{x}
when \code{x} is a poplin object.}

\item{xout}{Character specifying the name of data to store in \code{x} when
\code{x} is a poplin object.}
}
\value{
A poplin.tsne or \linkS4class{poplin} object with the same number of
rows as \code{ncol(x)} containing the dimension reduction result.
poplin.tsne is a matrix containing custom attributes used to summarize and
visualize the t-SNE result.
}
\description{
Apply t-SNE to a matrix or \linkS4class{poplin} object. This is an interface
to the \link[Rtsne]{Rtsne} function from the \pkg{Rtsne} package. t-SNE is
well-suited for visualizing high-dimensional data by giving each data point
a location in a two or three-dimensional map.
}
\examples{

data(faahko_poplin)

if (requireNamespace("Rtsne", quietly = TRUE)) {
  ## poplin object
  out <- reduce_tsne(faahko_poplin, xin = "knn_cyclic", xout = "tsne",
                     normalize = TRUE, perplexity = 3)
  summary(poplin_reduced(out, "tsne"))

  ## matrix
  m <- poplin_data(faahko_poplin, "knn_cyclic")
  out <- reduce_tsne(m, normalize = TRUE, perplexity = 3, ncomp = 3)
  summary(out)
}
}
\references{
L.J.P. van der Maaten and G.E. Hinton. Visualizing High-Dimensional Data
Using t-SNE. Journal of Machine Learning Research 9(Nov):2579-2605, 2008.

L.J.P. van der Maaten. Accelerating t-SNE using Tree-Based Algorithms.
Journal of Machine Learning Research 15(Oct):3221-3245, 2014.

Jesse H. Krijthe (2015). Rtsne: T-Distributed Stochastic Neighbor Embedding
using a Barnes-Hut Implementation, URL: https://github.com/jkrijthe/Rtsne
}
\seealso{
Other data reduction methods: 
\code{\link{poplin_reduce}()},
\code{\link{reduce_pca}()},
\code{\link{reduce_plsda}()}
}
\concept{data reduction methods}
