% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation-methods.R
\name{impute_knn}
\alias{impute_knn}
\alias{impute_knn,poplin-method}
\title{K-nearest neighbor (KNN) imputation}
\usage{
\S4method{impute_knn}{matrix}(x, by = c("feature", "sample"), ...)

\S4method{impute_knn}{poplin}(x, poplin_in, poplin_out, by = c("feature", "sample"), ...)
}
\arguments{
\item{x}{A matrix or \linkS4class{poplin} object.}

\item{by}{Imputation by k-nearest features or by k-nearest samples.}

\item{...}{Additional arguments passed to \link[VIM:kNN]{VIM::kNN}.}

\item{poplin_in}{Name of a data matrix to retrieve.}

\item{poplin_out}{Name of a data matrix to store.}
}
\value{
A matrix or \linkS4class{poplin} object of the same dimension as
\code{x} containing the imputed intensities.
}
\description{
Apply k-nearest neighbor (KNN) imputation to a matrix or
\linkS4class{poplin} object. This is an interface to the \link[VIM:kNN]{VIM::kNN} from the
\pkg{VIM} package. Since it is based on Gower's distance, standardization of
input data prior to KNN imputation would not affect the result.
}
\references{
Alexander Kowarik, Matthias Templ (2016). Imputation with the R Package VIM.
Journal of Statistical Software, 74(7), 1-16. doi:10.18637/jss.v074.i07

Gower, J. C. (1971). A General Coefficient of Similarity and Some of Its
Properties. Biometrics, 27(4), 857–871. https://doi.org/10.2307/2528823
}
\seealso{
Other imputation methods: 
\code{\link{impute_pca}()},
\code{\link{impute_randomforest}()},
\code{\link{impute_simple}()},
\code{\link{poplin_impute}()}
}
\concept{imputation methods}
