% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation-methods.R
\name{impute_pca}
\alias{impute_pca}
\alias{impute_pca,poplin-method}
\title{Principal component analysis (PCA) imputation}
\usage{
\S4method{impute_pca}{matrix}(x, type = c("bpca", "ppca", "nipals", "svdImpute"), ...)

\S4method{impute_pca}{poplin}(
  x,
  poplin_in,
  poplin_out,
  type = c("bpca", "ppca", "nipals", "svdImpute"),
  ...
)
}
\arguments{
\item{x}{A matrix or \linkS4class{poplin} object.}

\item{type}{A method for performing PCA.}

\item{...}{Additional arguments passed to \link[pcaMethods:pca]{pcaMethods::pca}.}

\item{poplin_in}{Name of a data matrix to retrieve.}

\item{poplin_out}{Name of a data matrix to store.}
}
\value{
A matrix or \linkS4class{poplin} object of the same dimension as
\code{x} containing the imputed intensities.
}
\description{
Apply PCA imputation to a matrix or \linkS4class{poplin} object. This is a
interface to the \link[pcaMethods:pca]{pcaMethods::pca} from the \pkg{pcaMethods} package. Here,
features are interpreted as variables and samples as observations.
Pre-processing of input (centering, scaling) may be necessary. See the
documentation of \link[pcaMethods:pca]{pcaMethods::pca} and \link[pcaMethods:prep]{pcaMethods::prep}. Note that the PCA
imputation could yield negative feature values that need to be
post-processed.
}
\references{
Stacklies, W., Redestig, H., Scholz, M., Walther, D. and Selbig, J.
pcaMethods -- a Bioconductor package providing PCA methods for incomplete
data. Bioinformatics, 2007, 23, 1164-1167
}
\seealso{
Other imputation methods: 
\code{\link{impute_knn}()},
\code{\link{impute_randomforest}()},
\code{\link{impute_simple}()},
\code{\link{poplin_impute}()}
}
\concept{imputation methods}
