% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation-methods.R
\name{impute_randomforest}
\alias{impute_randomforest}
\alias{impute_randomforest,matrix-method}
\alias{impute_randomforest,poplin-method}
\title{Random forest imputation}
\usage{
\S4method{impute_randomforest}{matrix}(x, ...)

\S4method{impute_randomforest}{poplin}(x, poplin_in, poplin_out, ...)
}
\arguments{
\item{x}{A matrix or \linkS4class{poplin} object.}

\item{...}{Additional arguments passed to \link[missForest:missForest]{missForest::missForest}.}

\item{poplin_in}{Name of a data matrix to retrieve.}

\item{poplin_out}{Name of a data matrix to store.}
}
\value{
A matrix or \linkS4class{poplin} object of the same dimension as
\code{x} containing the imputed intensities.
}
\description{
Apply random forest imputation to a matrix or \linkS4class{poplin} object.
This is an interface to the \link[missForest:missForest]{missForest::missForest} from the
\pkg{missForest} package. Since random forest is a tree-based method, it can
be performed with raw intensities - invariant to monotonic transformations.
}
\references{
Daniel J. Stekhoven (2013). missForest: Nonparametric Missing Value
Imputation using Random Forest. R package version 1.4.

Stekhoven D. J., & Buehlmann, P. (2012). MissForest - non-parametric missing
value imputation for mixed-type data. Bioinformatics, 28(1), 112-118.
}
\seealso{
Other imputation methods: 
\code{\link{impute_knn}()},
\code{\link{impute_pca}()},
\code{\link{impute_simple}()},
\code{\link{poplin_impute}()}
}
\concept{imputation methods}
