% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization-methods.R
\name{normalize_mean}
\alias{normalize_mean}
\alias{normalize_mean,poplin-method}
\title{Mean normalization}
\usage{
\S4method{normalize_mean}{matrix}(x, restrict = FALSE, rescale = FALSE)

\S4method{normalize_mean}{poplin}(x, poplin_in, poplin_out, restrict = FALSE, rescale = FALSE)
}
\arguments{
\item{x}{A matrix or \linkS4class{poplin} object.}

\item{restrict}{If \code{TRUE}, any feature with missing values is excluded
from the calculation of normalization factors.}

\item{rescale}{If \code{TRUE}, normalized intensities are multiplied by the
median of normalization factors to make look similar to their original
scales.}

\item{poplin_in}{Name of a data matrix to retrieve.}

\item{poplin_out}{Name of a data matrix to store.}
}
\value{
A matrix or \linkS4class{poplin} object of the same dimension as
\code{x} containing the normalized intensities.
}
\description{
Apply mean normalization to a matrix or \linkS4class{poplin} object. For
each sample, feature intensities are divided by its mean. The mean of
intensity values for individual samples will be one as a result.
}
\seealso{
Other normalization methods: 
\code{\link{normalize_cyclicloess}()},
\code{\link{normalize_mad}()},
\code{\link{normalize_median}()},
\code{\link{normalize_pqn}()},
\code{\link{normalize_scale}()},
\code{\link{normalize_sum}()},
\code{\link{normalize_vsn}()},
\code{\link{poplin_normalize}()}
}
\concept{normalization methods}
