% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization-methods.R
\name{normalize_vsn}
\alias{normalize_vsn}
\alias{normalize_vsn,matrix-method}
\alias{normalize_vsn,poplin-method}
\title{Variance stabilizing normalization (VSN)}
\usage{
\S4method{normalize_vsn}{matrix}(x, meanSdPlot = FALSE, ...)

\S4method{normalize_vsn}{poplin}(x, poplin_in, poplin_out, meanSdPlot = FALSE, ...)
}
\arguments{
\item{x}{A matrix or \linkS4class{poplin} object.}

\item{meanSdPlot}{If \code{TRUE}, the function shows \link[vsn:meanSdPlot]{vsn::meanSdPlot} to
visually verify whether there is a dependence of the standard deviation on
the mean.}

\item{...}{Additional arguments passed to \link[vsn:vsn2]{vsn::vsnMatrix}.}

\item{poplin_in}{Name of a data matrix to retrieve.}

\item{poplin_out}{Name of a data matrix to store.}
}
\value{
A matrix or \linkS4class{poplin} object of the same dimension as
\code{x} containing the normalized intensities.
}
\description{
Apply variance stabilizing normalization (VSN) to a matrix or
\linkS4class{poplin} object. This is an interface to the \link[vsn:vsn2]{vsn::vsnMatrix}
from the \pkg{vsn} package (see \link[vsn:vsn2]{vsn::vsn2} for help). The vsn produces
normalized intensities based on a glog (generalized logarithm) scale to base
2. See Huber et al. (2002) for details.
}
\references{
Huber W, von Heydebreck A, Sültmann H, Poustka A, Vingron M. Variance
stabilization applied to microarray data calibration and to the
quantification of differential expression. Bioinformatics. 2002;18 Suppl
1:S96-104. doi: 10.1093/bioinformatics/18.suppl_1.s96. PMID: 12169536.
}
\seealso{
Other normalization methods: 
\code{\link{normalize_cyclicloess}()},
\code{\link{normalize_mad}()},
\code{\link{normalize_mean}()},
\code{\link{normalize_median}()},
\code{\link{normalize_pqn}()},
\code{\link{normalize_scale}()},
\code{\link{normalize_sum}()},
\code{\link{poplin_normalize}()}
}
\concept{normalization methods}
