% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poplin-package.R
\docType{package}
\name{poplin-package}
\alias{poplin}
\alias{poplin-package}
\title{poplin: Reproducible analysis of metabolomics data}
\description{
This is a LC/MS data processing package.
}
\author{
\strong{Maintainer}: Jaehyun Joo \email{jaehyunjoo@outlook.com}

}
\keyword{internal}
