% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-boxplot.R
\name{poplin_boxplot}
\alias{poplin_boxplot}
\alias{poplin_boxplot.default}
\alias{poplin_boxplot.poplin}
\title{Box plot}
\usage{
\method{poplin_boxplot}{default}(
  x,
  group,
  pre_log2 = FALSE,
  violin = FALSE,
  ylab = "Intensity",
  ...
)

\method{poplin_boxplot}{poplin}(
  x,
  poplin_in,
  group,
  pre_log2 = FALSE,
  violin = FALSE,
  ylab = "Intensity",
  ...
)
}
\arguments{
\item{x}{A matrix or \linkS4class{poplin} object.}

\item{group}{A discrete variable to visualize the grouping structure.}

\item{pre_log2}{If \code{TRUE}, feature intensities are log2-transformed
before plotting.}

\item{violin}{If \code{TRUE}, a violin plot is drawn instead of the boxplot.}

\item{ylab}{The title of y-axis of the plot.}

\item{...}{Reserved for future use.}

\item{poplin_in}{Name of a data matrix to retrieve.}
}
\value{
A ggplot object.
}
\description{
Produce a box-and-whisker plot of the feature intensity values.
}
