% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-corplot.R
\name{poplin_corplot}
\alias{poplin_corplot}
\alias{poplin_corplot.default}
\alias{poplin_corplot.poplin}
\title{Correlation plot}
\usage{
\method{poplin_corplot}{default}(
  x,
  unit = c("sample", "feature"),
  use = c("everything", "all.obs", "complete.obs", "na.or.complete",
    "pairwise.complete.obs"),
  method = c("pearson", "kendall", "spearman"),
  widths = NULL,
  heights = NULL,
  colors = viridis::viridis(n = 256, alpha = 1, begin = 0, end = 1, option = "viridis"),
  label = FALSE,
  digits = 2,
  grid_gap = 0,
  hide_colorbar = FALSE,
  showticklabels = c(TRUE, TRUE),
  row_dend_left = FALSE,
  ...
)

\method{poplin_corplot}{poplin}(
  x,
  xin,
  unit = c("sample", "feature"),
  use = c("everything", "all.obs", "complete.obs", "na.or.complete",
    "pairwise.complete.obs"),
  method = c("pearson", "kendall", "spearman"),
  widths = NULL,
  heights = NULL,
  colors = viridis::viridis(n = 256, alpha = 1, begin = 0, end = 1, option = "viridis"),
  label = FALSE,
  digits = 2,
  grid_gap = 0,
  hide_colorbar = FALSE,
  showticklabels = c(TRUE, TRUE),
  row_dend_left = FALSE,
  ...
)
}
\arguments{
\item{x}{a matrix or \linkS4class{poplin} object.}

\item{unit}{character specifying whether a correlation matrix is
computed based on samples or features.}

\item{use}{the method to compute correlations in the presence of missing
values. Refer to \code{?cor} for details.}

\item{method}{character specifying which correlation coefficient is
to be computed. Refer to \code{?cor} for details.}

\item{widths}{relative widths of heatmap and dendrogram.}

\item{heights}{relative heights of heatmap and dendrogram.}

\item{colors}{a vector of colors for heatmap.}

\item{label}{logical controlling whether cell values are shown.}

\item{digits}{the desired number of digits when \code{label = TRUE}.}

\item{grid_gap}{numeric specifying the gap between cells.}

\item{hide_colorbar}{logical controlling whether the color bar (legend) is
hidden.}

\item{showticklabels}{a logical vector of length 2 (x-axis, y-axis). If
\code{FALSE}, the ticks are removed from the sides of the plot.}

\item{row_dend_left}{logical controlling whether the row dendrogram is
placed on the left on the plot.}

\item{...}{additional arguments passed to \link[heatmaply:heatmaply]{heatmaply::heatmaply}.}

\item{xin}{character specifying the name of data to retrieve from \code{x}
when \code{x} is a poplin object.}
}
\value{
gtable of aligned plots.
}
\description{
Visualize correlations between samples or features. All values in a
correlation matrix are visualized by rectangles.
}
\examples{

## poplin object
poplin_corplot(faahko_poplin, xin = "knn_cyclic")

## matrix
m <- poplin_data(faahko_poplin, "knn_cyclic")
poplin_corplot(m, label = TRUE)
}
