% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation-methods.R
\name{poplin_impute}
\alias{poplin_impute}
\alias{poplin_impute,matrix-method}
\alias{poplin_impute,poplin-method}
\title{Imputation methods}
\usage{
\S4method{poplin_impute}{matrix}(x, method = c("knn", "randomforest", "pca", "univariate"), ...)

\S4method{poplin_impute}{poplin}(
  x,
  method = c("knn", "randomforest", "pca", "univariate"),
  xin,
  xout,
  ...
)
}
\arguments{
\item{x}{A matrix or \linkS4class{poplin} object.}

\item{method}{The imputation method to be used, defaulting to "knn".}

\item{...}{Arguments passed to a specific imputation method.}

\item{xin}{Character specifying the name of data to retrieve from \code{x}
when \code{x} is a poplin object.}

\item{xout}{Character specifying the name of data to store in \code{x} when
\code{x} is a poplin object.}
}
\value{
A matrix or \linkS4class{poplin} object of the same dimension as
\code{x} containing the imputed intensities.
}
\description{
Missing values are frequently found in metabolomics data. The \pkg{poplin}
package provides a few options to handle them.
\link{poplin_impute} is a wrapper for the following set of functions:
\describe{
\item{\code{\link{impute_randomforest}}:}{
random forest imputation
}
\item{\code{\link{impute_knn}}:}{
k-nearest neighbor (KNN) imputation
}
\item{\code{\link{impute_pca}}:}{
principal component analysis (PCA) imputation
}
\item{\code{\link{impute_univariate}}:}{
univariate imputation (e.g., half-minimum, mean, median)
}
}
}
\examples{

data(faahko_poplin)

if (requireNamespace("VIM", quietly = TRUE)) {
 ## poplin object
 poplin_impute(faahko_poplin, method = "knn", xin = "raw", xout = "knn")

 ## matrix
 m <- poplin_raw(faahko_poplin, "raw")
 poplin_impute(m, method = "knn")
}
}
\seealso{
Other imputation methods: 
\code{\link{impute_knn}()},
\code{\link{impute_pca}()},
\code{\link{impute_randomforest}()},
\code{\link{impute_univariate}()}
}
\concept{imputation methods}
