% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation-methods.R
\name{poplin_impute}
\alias{poplin_impute}
\alias{poplin_impute,poplin-method}
\title{Imputation methods}
\usage{
\S4method{poplin_impute}{matrix}(x, method = c("knn", "pca", "randomforest", "simple"), ...)

\S4method{poplin_impute}{poplin}(
  x,
  method = c("knn", "pca", "randomforest", "simple"),
  poplin_in,
  poplin_out,
  ...
)
}
\arguments{
\item{x}{A matrix or \linkS4class{poplin} object.}

\item{method}{A imputation method. Default is 'pqn'.}

\item{...}{Arguments passed to a specific imputation method.}

\item{poplin_in}{Name of a data matrix to retrieve.}

\item{poplin_out}{Name of a data matrix to store.}
}
\value{
A matrix or \linkS4class{poplin} object of the same dimension as
\code{x} containing the imputed intensities.
}
\description{
Missing values are frequently found in metabolomics data. The \pkg{poplin}
package provides a few options to handle them.
\link{poplin_impute} is a wrapper for the following set of functions:
\describe{
\item{\code{\link{impute_knn}}:}{
k-nearest neighbor (KNN) imputation
}
\item{\code{\link{impute_pca}}:}{
principal component analysis (PCA) imputation
}
\item{\code{\link{impute_randomforest}}:}{
random forest imputation
}
\item{\code{\link{impute_simple}}:}{
simple univariate imputation (e.g., half-minimum, mean, median)
}
}
}
\seealso{
Other imputation methods: 
\code{\link{impute_knn}()},
\code{\link{impute_pca}()},
\code{\link{impute_randomforest}()},
\code{\link{impute_simple}()}
}
\concept{imputation methods}
