% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-naplot.R
\name{poplin_naplot}
\alias{poplin_naplot}
\alias{poplin_naplot.default}
\alias{poplin_naplot.poplin}
\title{Missing value plot}
\usage{
\method{poplin_naplot}{default}(
  x,
  widths = NULL,
  heights = NULL,
  colors = viridis::viridis(2),
  label = FALSE,
  digits = 2,
  grid_gap = 1,
  hide_colorbar = TRUE,
  showticklabels = c(TRUE, FALSE),
  row_dend_left = FALSE,
  ...
)

\method{poplin_naplot}{poplin}(
  x,
  xin,
  widths = NULL,
  heights = NULL,
  colors = viridis::viridis(2),
  label = FALSE,
  digits = 2,
  grid_gap = 1,
  hide_colorbar = TRUE,
  showticklabels = c(TRUE, FALSE),
  row_dend_left = FALSE,
  ...
)
}
\arguments{
\item{x}{A matrix or \linkS4class{poplin} object.}

\item{widths}{Relative widths of heatmap and dendrogram.}

\item{heights}{Relative heights of heatmap and dendrogram.}

\item{colors}{A vector of colors for heatmap.}

\item{label}{Logical controlling whether cell values are shown.}

\item{digits}{The desired number of digits when \code{label = TRUE}.}

\item{grid_gap}{Numeric specifying the gap between cells.}

\item{hide_colorbar}{Logical controlling whether the color bar (legend) is
hidden.}

\item{showticklabels}{A logical vector of length 2 (x-axis, y-axis). If
\code{FALSE}, the ticks are removed from the sides of the plot.}

\item{row_dend_left}{Logical controlling whether the row dendrogram is
placed on the left on the plot.}

\item{...}{Additional arguments passed to \link[heatmaply]{heatmaply}
(\pkg{heatmaply} package).}

\item{xin}{Character specifying the name of data to retrieve from \code{x}
when \code{x} is a poplin object.}
}
\value{
A gtable of aligned plots
}
\description{
Visualize data for exploring missing value (NA) patterns. All values in a
data matrix are recoded (1: missing; 0: non-missing) and visualized by
rectangles.
}
\examples{

data(faahko_poplin)

## poplin object
poplin_naplot(faahko_poplin, xin = "raw")

## matrix
m <- poplin_raw(faahko_poplin, "raw")
poplin_naplot(m)
}
