% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization-methods.R
\name{poplin_normalize}
\alias{poplin_normalize}
\alias{poplin_normalize,poplin-method}
\title{Normalization methods}
\usage{
\S4method{poplin_normalize}{matrix}(
  x,
  method = c("pqn", "sum", "mean", "median", "mad", "euclidean", "cyclicloess", "vsn",
    "scale"),
  ...
)

\S4method{poplin_normalize}{poplin}(
  x,
  method = c("pqn", "sum", "mean", "median", "mad", "euclidean", "cyclicloess", "vsn",
    "scale"),
  poplin_in,
  poplin_out,
  ...
)
}
\arguments{
\item{x}{A matrix or \linkS4class{poplin} object.}

\item{method}{A normalization method. Default is 'pqn'.}

\item{...}{Argument passed to a specific normalization method.}

\item{poplin_in}{Name of a data matrix to retrieve.}

\item{poplin_out}{Name of a data matrix to store.}
}
\value{
A matrix or \linkS4class{poplin} object of the same dimension as
\code{x} containing the normalized intensities.
}
\description{
Data normalization aims to reduce unwanted variations. The \pkg{poplin}
package provides a few data-driven normalization methods.
\link{poplin_normalize} is a wrapper for the following set of functions:
\describe{
\item{\code{\link{poplin_normalize_sum}}:}{
sum normalization (TIC normalization)
}
\item{\code{\link{poplin_normalize_mean}}:}{
mean normalization
}
\item{\code{\link{poplin_normalize_median}}:}{
median normalization
}
\item{\code{\link{poplin_normalize_mad}}:}{
MAD (median absolute deviation) normalization
}
\item{\code{\link{poplin_normalize_cyclicloess}}:}{
cyclic LOESS normalization
}
\item{\code{\link{poplin_normalize_pqn}}:}{
PQN (probabilistic quotient normalization)
}
\item{\code{\link{poplin_normalize_vsn}}:}{
VSN (variance stabilizing normalization)
}
\item{\code{\link{poplin_normalize_scale}}:}{
feature-based scaling (e.g., auto, range, pareto, vast, level)
}
}
}
\seealso{
Other normalization methods: 
\code{\link{poplin_normalize_cyclicloess}()},
\code{\link{poplin_normalize_mad}()},
\code{\link{poplin_normalize_mean}()},
\code{\link{poplin_normalize_median}()},
\code{\link{poplin_normalize_pqn}()},
\code{\link{poplin_normalize_scale}()},
\code{\link{poplin_normalize_sum}()},
\code{\link{poplin_normalize_vsn}()}
}
\concept{normalization methods}
