% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction-methods.R
\name{poplin_reduce}
\alias{poplin_reduce}
\alias{poplin_reduce,matrix-method}
\alias{poplin_reduce,poplin-method}
\title{Dimension reduction methods}
\usage{
\S4method{poplin_reduce}{matrix}(x, method = c("pca", "tsne", "plsda"), y, ncomp = 2, ...)

\S4method{poplin_reduce}{poplin}(
  x,
  method = c("pca", "tsne", "plsda"),
  xin,
  xout,
  y,
  ncomp = 2,
  ...
)
}
\arguments{
\item{x}{A matrix or \linkS4class{poplin} object.}

\item{method}{The dimension reduction method to be used, defaulting to
"pca".}

\item{y}{A factor vector for discrete outcome required for PLS-DA. Ignored
otherwise.}

\item{ncomp}{Output dimensionality.}

\item{...}{Argument passed to a specific dimension reduction method.}

\item{xin}{Character specifying the name of data to retrieve from \code{x}
when \code{x} is a poplin object.}

\item{xout}{character specifying the name of data to store in \code{x} when
\code{x} is a poplin object.}
}
\value{
A matrix or \linkS4class{poplin} object with the same number of
rows as \code{ncol(x)} containing the dimension reduction result.
}
\description{
In metabolomics, dimension reduction methods are often used for modeling
and visualization.
\link{poplin_reduce} is a wrapper for the following set of functions:
\describe{
\item{\code{\link{reduce_pca}}:}{
principal component analysis (PCA)
}
\item{\code{\link{reduce_plsda}}:}{
partial least squares-discriminant analysis (PLS-DA)
}
\item{\code{\link{reduce_tsne}}:}{
t-distributed stochastic neighbor embedding
}
}
}
\examples{

data(faahko_poplin)

## poplin object
out <- poplin_reduce(faahko_poplin, method = "pca",
                     xin = "knn_cyclic", xout = "pca")
summary(poplin_reduced(out, "pca"))

## matrix
m <- poplin_data(faahko_poplin, "knn_cyclic")
poplin_reduce(m, method = "pca")
summary(out)
}
\seealso{
Other data reduction methods: 
\code{\link{reduce_pca}()},
\code{\link{reduce_plsda}()},
\code{\link{reduce_tsne}()}
}
\concept{data reduction methods}
