% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction-methods.R
\name{poplin_reduce}
\alias{poplin_reduce}
\alias{poplin_reduce,poplin-method}
\title{Dimension reduction methods}
\usage{
\S4method{poplin_reduce}{matrix}(x, method = c("pca", "tsne", "plsda"), y, ncomp = 2, ...)

\S4method{poplin_reduce}{poplin}(
  x,
  method = c("pca", "tsne", "plsda"),
  poplin_in,
  poplin_out,
  y,
  ncomp = 2,
  ...
)
}
\arguments{
\item{x}{A matrix or \linkS4class{poplin} object.}

\item{method}{A dimension reduction method. Default is 'pca'.}

\item{y}{A factor vector for discrete outcome required for PLS-DA. Ignored
otherwise.}

\item{ncomp}{Output dimensionality.}

\item{...}{Argument passed to a specific dimension reduction method.}

\item{poplin_in}{Name of a data matrix to retrieve.}

\item{poplin_out}{Name of a data matrix to store.}
}
\value{
A matrix or \linkS4class{poplin} object with the same number of rows
as \code{ncol(x)} containing the dimension reduction result.
}
\description{
In metabolomics, dimension reduction methods are often used for modeling
and visualization.
\link{poplin_reduce} is a wrapper for the following set of functions:
\describe{
\item{\code{\link{poplin_reduce_pca}}:}{
principal component analysis (PCA)
}
\item{\code{\link{poplin_reduce_plsda}}:}{
partial least squares-discriminant analysis (PLS-DA)
}
\item{\code{\link{poplin_reduce_tsne}}:}{
t-distributed stochastic neighbor embedding
}
}
}
\seealso{
Other data reduction methods: 
\code{\link{poplin_reduce_pca}()},
\code{\link{poplin_reduce_plsda}()},
\code{\link{poplin_reduce_tsne}()}
}
\concept{data reduction methods}
