% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-scoreplot.R
\name{poplin_scoreplot}
\alias{poplin_scoreplot}
\alias{poplin_scoreplot.default}
\alias{poplin_scoreplot.poplin.pca}
\alias{poplin_scoreplot.poplin.plsda}
\alias{poplin_scoreplot.poplin}
\title{Score plot of dimension-reduced data}
\usage{
\method{poplin_scoreplot}{default}(
  x,
  comp = c(1, 2),
  group,
  group_col = NULL,
  point_size = 1.5,
  point_shape_by_group = FALSE,
  label = FALSE,
  label_size = 3.88,
  label_subset = NULL,
  ellipse = FALSE,
  xlab = NULL,
  ylab = NULL,
  title = NULL,
  legend = TRUE,
  ...
)

\method{poplin_scoreplot}{poplin.pca}(
  x,
  comp = c(1, 2),
  group,
  group_col = NULL,
  label = FALSE,
  xlab = NULL,
  ylab = NULL,
  ...
)

\method{poplin_scoreplot}{poplin.plsda}(
  x,
  comp = c(1, 2),
  group = attr(x, "Y.observed"),
  group_col = NULL,
  label = FALSE,
  xlab = NULL,
  ylab = NULL,
  ...
)

\method{poplin_scoreplot}{poplin}(
  x,
  xin,
  comp = c(1, 2),
  group,
  group_col = NULL,
  label = FALSE,
  xlab = NULL,
  ylab = NULL,
  ...
)
}
\arguments{
\item{x}{A dimension-reduced data matrix produced by \link{poplin_reduce} or
\linkS4class{poplin} object containing dimension-reduced data.}

\item{comp}{A numeric vector of length 2 indicating two components to plot.}

\item{group}{A discrete variable to visualize the grouping structure.}

\item{group_col}{A vector of colors with the same length of unique values in
\code{group}.}

\item{point_size}{Numeric controlling the size of points.}

\item{point_shape_by_group}{Logical controlling whether each group have
different shapes of data points. Also can be a numeric vector with the
same length of unique values in \code{group} to manually set point shapes.}

\item{label}{Logical controlling whether score labels are shown instead of
points.}

\item{label_size}{Numeric controlling the size of labels.}

\item{label_subset}{A character vector specifying a subset of score labels
to display.}

\item{ellipse}{Logical controlling whether data ellipses are shown using
the \link[ggplot2]{stat_ellipse} function from the \pkg{ggplot2} package.}

\item{xlab}{The title of x-axis of the plot.}

\item{ylab}{The title of y-axis of the plot.}

\item{title}{The main title of the plot.}

\item{legend}{Logical controlling whether the plot legend is shown.}

\item{...}{Arguments passed to the default method.}

\item{xin}{Character specifying the name of data to retrieve from \code{x}
when \code{x} is a poplin object}
}
\value{
A ggplot object.
}
\description{
Visualize the data onto a lower-dimensional space using the \link{poplin_reduce}
output.
}
\examples{

data(faahko_poplin)

## sample group variable
group <- colData(faahko_poplin)$sample_group

## poplin object
poplin_scoreplot(faahko_poplin, xin = "pca", group = group)

## matrix
m <- poplin_reduced(faahko_poplin, "pca")
poplin_scoreplot(m, group = group, label = TRUE)
}
\seealso{
\link{poplin_reduce}, \link{poplin_biplot}.
}
