% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction-methods.R
\name{reduce_pca}
\alias{reduce_pca}
\alias{reduce_pca,matrix-method}
\alias{reduce_pca,poplin-method}
\title{Principal component analysis (PCA)}
\usage{
\S4method{reduce_pca}{matrix}(x, ncomp = 2, center = TRUE, scale = FALSE, ...)

\S4method{reduce_pca}{poplin}(
  x,
  poplin_in,
  poplin_out,
  ncomp = 2,
  center = TRUE,
  scale = FALSE,
  ...
)
}
\arguments{
\item{x}{A matrix or \linkS4class{poplin} object.}

\item{ncomp}{Output dimensionality.}

\item{center}{A logical indicating mean-centering prior to PCA.}

\item{scale}{A logical indicating unit variance scaling prior to PCA.}

\item{...}{Additional arguments passed to \link[pcaMethods:nipalsPca]{pcaMethods::nipalsPca}.}

\item{poplin_in}{Name of a data matrix to retrieve.}

\item{poplin_out}{Name of a data matrix to store.}
}
\value{
A poplin.pca matrix or \linkS4class{poplin} object with the same
number of rows as \code{ncol(x)} containing the dimension reduction
result.
}
\description{
Apply PCA to a matrix or \linkS4class{poplin} object. For the data without
missing values, PCA is performed via singular value decomposition.
Otherwise, PCA is performed using the non-linear iterative partial least
squares (NIPALS) algorithm via \link[pcaMethods:nipalsPca]{pcaMethods::nipalsPca} from the
\pkg{pcaMethods} package. NIPALS PCA can handle a small amount of missing
values.
}
\references{
Wold, H. (1966). Estimation of principal components and related models by
iterative least squares. In P. R. Krishnajah (Ed.), Multivariate analysis
(pp. 391-420). NewYork: Academic Press.
}
\seealso{
Other data reduction methods: 
\code{\link{poplin_reduce}()},
\code{\link{reduce_plsda}()},
\code{\link{reduce_tsne}()}
}
\concept{data reduction methods}
