% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction-methods.R
\name{reduce_plsda}
\alias{reduce_plsda}
\alias{reduce_plsda,matrix-method}
\alias{reduce_plsda,poplin-method}
\title{Partial least squares-discriminant analysis (PLS-DA)}
\usage{
\S4method{reduce_plsda}{matrix}(x, y, ncomp = 2, center = TRUE, scale = FALSE, ...)

\S4method{reduce_plsda}{poplin}(
  x,
  poplin_in,
  poplin_out,
  y,
  ncomp = 2,
  center = TRUE,
  scale = FALSE,
  ...
)
}
\arguments{
\item{x}{A matrix or \linkS4class{poplin} object.}

\item{y}{A factor vector for discrete outcome.}

\item{ncomp}{Output dimensionality.}

\item{center}{A logical indicating mean-centering prior to PLS-DA.}

\item{scale}{A logical indicating unit variance scaling prior to PLS-DA.}

\item{...}{Additional arguments passed to \link[pls:mvr]{pls::plsr}.}

\item{poplin_in}{Name of a data matrix to retrieve.}

\item{poplin_out}{Name of a data matrix to store.}
}
\value{
A poplin.plsda matrix or \linkS4class{poplin} object with the same
number of rows as \code{ncol(x)} containing the dimension reduction
result.
}
\description{
Apply PLS-DA to a matrix or \linkS4class{poplin} object. It performs
standard PLS for classification using \link[pls:mvr]{pls::plsr}. If the \pkg{pls} is not
installed, this function will stop with a note about install the package.
}
\references{
Kristian Hovde Liland, Bjørn-Helge Mevik and Ron Wehrens (2021). pls:
Partial Least Squares and Principal Component Regression. R package version
2.8-0. https://CRAN.R-project.org/package=pls
}
\seealso{
Other data reduction methods: 
\code{\link{poplin_reduce}()},
\code{\link{reduce_pca}()},
\code{\link{reduce_tsne}()}
}
\concept{data reduction methods}
