% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction-methods.R
\name{reduce_tsne}
\alias{reduce_tsne}
\alias{reduce_tsne,matrix-method}
\alias{reduce_tsne,poplin-method}
\title{t-distributed stochastic neighbor embedding (t-SNE)}
\usage{
\S4method{reduce_tsne}{matrix}(x, ncomp = 2, normalize = TRUE, ...)

\S4method{reduce_tsne}{poplin}(x, poplin_in, poplin_out, ncomp = 2, normalize = TRUE, ...)
}
\arguments{
\item{x}{A matrix or \linkS4class{poplin} object.}

\item{ncomp}{Number of components to calculate.}

\item{normalize}{if \code{TRUE}, an input matrix is mean-centered and scaled
so that the largest absolute of the centered matrix is equal to unity. See
\link[Rtsne:normalize_input]{Rtsne::normalize_input} for details.}

\item{...}{Additional arguments passed to \link[Rtsne:Rtsne]{Rtsne::Rtsne}.}

\item{poplin_in}{Name of a data matrix to retrieve.}

\item{poplin_out}{Name of a data matrix to store.}
}
\value{
A poplin.tsne matrix or \linkS4class{poplin} object with the same
number of rows as \code{ncol(x)} containing the dimension reduction
result.
}
\description{
Apply t-SNE to a matrix or \linkS4class{poplin} object. This is an interface
to the \link[Rtsne:Rtsne]{Rtsne::Rtsne} from the \pkg{Rtsne} package. t-SNE is well-suited for
visualizing high-dimensional data by giving each data point a location in a
two or three-dimensional map.
}
\references{
L.J.P. van der Maaten and G.E. Hinton. Visualizing High-Dimensional Data
Using t-SNE. Journal of Machine Learning Research 9(Nov):2579-2605, 2008.

L.J.P. van der Maaten. Accelerating t-SNE using Tree-Based Algorithms.
Journal of Machine Learning Research 15(Oct):3221-3245, 2014.

Jesse H. Krijthe (2015). Rtsne: T-Distributed Stochastic Neighbor Embedding
using a Barnes-Hut Implementation, URL: https://github.com/jkrijthe/Rtsne
}
\seealso{
Other data reduction methods: 
\code{\link{poplin_reduce}()},
\code{\link{reduce_pca}()},
\code{\link{reduce_plsda}()}
}
\concept{data reduction methods}
